app/src/main/java/com/example/library2/pluralsight/junit/course1/list/ListViewModel.kt
package com.example.library2.pluralsight.junit.course1.list

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import com.example.library2.pluralsight.junit.course1.data.Todo
import com.example.library2.pluralsight.junit.course1.data.TodoRepository

class ListViewModel(
    private val todoRepository: TodoRepository
) : ViewModel() {

    val allTodos: LiveData<List<Todo>> = todoRepository.getAllTodos()
    val upcomingTodosCount: LiveData<Int> = todoRepository.getUpcomingTodosCount()

    fun toggleTodo(id: String) {
        todoRepository.toggleTodo(id)
    }

}